+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.0.7
 
 - Only play sound effects for the current user (and don't push to other clients).
-- Add Virtua Boi red color scheme.
+- Add Virtua Boi and p0w3rsh3ll color schemes.
 
 # 1.0.6